requirements.txt
pandas==2.2.3
numpy==2.2.4
scipy==1.15.2
scikit-learn==1.5.2
jpype1==1.5.1
matplotlib==3.10.3
seaborn==0.13.2
tqdm==4.66.4
jupyter==1.1.0
nbformat==5.10.4
pyyaml==6.0.2
statsmodels==0.14.4
